feat: toggle color mode ubication refactor
